a bigger refactor; wip
