+++ src/mcp_runtime_server/sandboxes/platforms.py
@@ -1,0 +1,1 @@
+"""Platform detection and mapping."""
