+++ src/mcp_runtime_server/types.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
-from typing import Dict, Any, Optional, List
+from typing import Dict, Any, Optional, List, NamedTuple
 from tempfile import TemporaryDirectory
 
 
@@ -36,8 +36,32 @@ class PackageManager(str, Enum):
 
 
 @dataclass(frozen=True)
+class PlatformInfo:
+    """Platform information."""
+
+    os_name: str
+    arch: str
+    format: str
+    node_platform: str
+    bun_platform: str
+    uv_platform: str
+
+
+class PlatformMapping(NamedTuple):
+    """Platform-specific values."""
+
+    node: str
+    bun: str
+    uv: str
+    archive_format: str
+    platform_template: str
+    binary_location: str
+
+
+@dataclass(frozen=True)
 class RuntimeConfig:
     """Runtime configuration details."""
+
     config_files: List[str]  # Files that indicate this runtime
     package_manager: PackageManager  # Default package manager
     env_setup: Dict[str, str]  # Base environment variables
@@ -57,23 +81,25 @@ class Sandbox:
     root: Path
     work_dir: Path
     bin_dir: Path
+    temp_dir: TemporaryDirectory
     env_vars: Dict[str, str]
 
 
 @dataclass(frozen=True)
 class Environment:
     """Runtime environment instance."""
+
     id: str
     runtime: Runtime
     created_at: datetime
     env_vars: Dict[str, str]
     sandbox: Sandbox
-    tempdir: TemporaryDirectory
 
 
 @dataclass
 class TestCase:
     """Test case execution result."""
+
     name: str
     status: str
     output: List[str]
@@ -84,6 +110,7 @@ class TestCase:
 @dataclass
 class RunTestResult:
     """Results from a test framework run."""
+
     success: bool
     framework: str
     passed: Optional[int] = None
