+++ src/mcp_runtime_server/utils.py
@@ -1,0 +1,105 @@
+import aiohttp
+import zipfile
+import tarfile
+from typing import List, Union
+from pathlib import Path
+from mcp_runtime_server.logging import get_logger
+
+logger = get_logger(__name__)
+
+
+async def download_url(url: str, dest: Path) -> None:
+    """Download a binary file with checksum verification."""
+    try:
+        async with aiohttp.ClientSession() as session:
+            async with session.get(url) as response:
+                if response.status != 200:
+                    raise RuntimeError(f"Download failed with status {response.status}")
+
+                with open(dest, "wb") as f:
+                    while chunk := await response.content.read(8192):
+                        f.write(chunk)
+
+    except Exception as e:
+        if dest.exists():
+            dest.unlink()
+        raise RuntimeError(f"Failed to download url: {e}")
+
+
+async def verify_checksum(file_path: Path, checksum_url: str) -> bool:
+    """Verify the checksum of a downloaded binary."""
+    try:
+        async with aiohttp.ClientSession() as session:
+            async with session.get(checksum_url) as response:
+                response.raise_for_status()
+                checksums = await response.text()
+
+                for line in checksums.splitlines():
+                    if file_path.name in line:
+                        expected_checksum, _ = line.split()
+                        import hashlib
+
+                        with open(file_path, "rb") as f:
+                            actual_checksum = hashlib.sha256(f.read()).hexdigest()
+                            return actual_checksum == expected_checksum
+
+        return False
+    except Exception as e:
+        logger.error(
+            {
+                "event": "checksum_verification_failed",
+                "error": str(e),
+                "file": str(file_path),
+                "checksum_url": checksum_url,
+            }
+        )
+        return False
+
+
+def get_archive_files(
+    archive: Union[zipfile.ZipFile, tarfile.TarFile], format: str
+) -> List[str]:
+    """Get list of files from archive handling different archive types."""
+    try:
+        if isinstance(archive, zipfile.ZipFile):
+            return archive.namelist()
+        else:  # tarfile.TarFile
+            return archive.getnames()
+    except Exception as e:
+        logger.error(
+            {"event": "list_archive_failed", "format": format, "error": str(e)}
+        )
+        raise ValueError(f"Failed to read {format} archive") from e
+
+
+def extract_archive(archive_path: Path, dest_dir: Path) -> Path:
+    """Extract binary from archive with flexible support."""
+    format = (
+        "".join(archive_path.suffixes[-2:])
+        if len(archive_path.suffixes) > 1
+        else archive_path.suffix
+    )
+
+    archive_handlers = {
+        ".zip": zipfile.ZipFile,
+        ".tar.gz": tarfile.open,
+        ".tgz": tarfile.open,
+    }
+
+    handler = archive_handlers.get(format)
+    if not handler:
+        raise ValueError(f"Unsupported archive format: {format}")
+
+    with handler(archive_path) as archive:
+        all_files = get_archive_files(archive, format)
+        archive.extract(all_files, dest_dir)
+
+        logger.info(
+            {
+                "event": "archive_extracted",
+                "archive": str(archive_path),
+                "extracted_to": str(dest_dir),
+            }
+        )
+
+        return dest_dir
